modified:   src/prologue.md Since the article  has reached the core completion, I am planning to maintain a version history hereon. The idea is to track how interact with this knowledge artifact hereon.
Up until now, there were many things shaping up simultaneously. The
storyline , the style, the format and the depth. As such it was
difficult to keep a track of evolution. Plus I was not sure if the
article had make it to the final cut of the book.

I think, now I have a good handle on all these aspects. I still need to add final pieces of
the story but those are well formed in my mind.

-----

The plan is to have at least one commit - preferably the last commit of
the day, capture the spirit of all the changes for the day. I want to
reach a cadence level where I write the motivation for the changes for
every single commit but sometimes that is just too much and don't want
to beat me down for that.

---

Version history, along with comments are the knowledge part of a piece
of text. Text in itself - in the final shape - is just information. The
thing that should have a value is the version history and commentary ;
not the final copy.

---
Changes of the day

Started reading the article from the first line. It is a fast and easy
read but felt like the starting needs to be more immersive for the
reader. So the focus today was to fix the opening ten pages. In the
morning session, I ordered the flow. Dropped few paras. Made the
arguments bit more sharp and language bit more conducive. In the evening
session , started writing a new paragraph to ease the reader in. A
sensational start maybe.

---

Key concepts added
- In the days if spoken word, the idea of time keeping on scriptures
  were pretty insignificant. That is the reason , texts of early
  scriptures have no date.
- Astrology needed time keeping as well charting mechanism much beyond
  the spoken word. Even beyond the written word.

---

Regular typos and Language tool grammar objections.